Remove unneeded pods, update balance widget
